wuber/requirements.txt

django==5.1.4


djangorestframework==3.15.1


djangorestframework-simplejwt==5.3.1






